+++ Package.swift
@@ -10,7 +10,7 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://source.skip.tools/skip.git", from: "1.8.2"),
-        .package(url: "https://source.skip.tools/skip-ui.git", from: "1.50.0")
+        .package(url: "https://source.skip.tools/skip-ui.git", from: "1.51.0")
     ],
     targets: [
         .target(name: "SkipWeb", dependencies: [.product(name: "SkipUI", package: "skip-ui")], resources: [.process("Resources")], plugins: [.plugin(name: "skipstone", package: "skip")]),